% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dca.R
\name{cols_n_rows_labels}
\alias{cols_n_rows_labels}
\title{Give a column and row name to all DCA components in a list}
\usage{
cols_n_rows_labels(
  data,
  n_fd,
  n_sd,
  periods,
  fd_names,
  sd_names,
  periods_names
)
}
\arguments{
\item{data}{DCA components list}

\item{n_fd}{n_fd from dca() function}

\item{n_sd}{n_sd from dca() function}

\item{periods}{periods from dca() function}

\item{fd_names}{fd_names from dca() function}

\item{sd_names}{sd_names from dca() function}

\item{periods_names}{periods_names from dca() function}
}
\value{
A list containing all DCA components correctly named
}
\description{
This function attribute a named number from 1:size of columns or rows ,or
attribute the names given by fd_names, sd_names and periods_names, at each
matrix component.
}
