% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dca.R
\name{cols_n_rows_names}
\alias{cols_n_rows_names}
\title{Give a column and row name to all DCA components in a list}
\usage{
cols_n_rows_names(data)
}
\arguments{
\item{data}{DCA components list}
}
\value{
A list containing all DCA components correctly named
}
\description{
This function attribute a named number from 1:size of columns or rows at each
matrix component.
}
