% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dca.R
\name{dca}
\alias{dca}
\title{Runs the DCA method}
\usage{
dca(data, n_fd, n_sd, periods)
}
\arguments{
\item{data}{Suitable dataset object}

\item{n_fd}{Number of indexers in first dimension}

\item{n_sd}{Number of indexers in second dimension}

\item{period}{Number of periods of the Concentration Index}
}
\value{
A list containing all DCA components
}
\description{
This function runs the entire Data Concentration Analysis to the desired dataset.
A proper explication of the methodology will be soon added do GitHub page.
}
