% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dca.R
\name{dca}
\alias{dca}
\title{Runs the DCA method}
\usage{
dca(
  data,
  n_fd,
  n_sd,
  periods,
  fd_names = NULL,
  sd_names = NULL,
  periods_names = NULL
)
}
\arguments{
\item{data}{Suitable dataset object}

\item{n_fd}{Number of indexers in first dimension}

\item{n_sd}{Number of indexers in second dimension}

\item{periods}{Number of periods of the Concentration Index}

\item{fd_names}{A vector containing first dimension names. Size must be the same as n_fd.}

\item{sd_names}{A vector containing second dimension names. Size must be the same as n_sd.}

\item{periods_names}{A vector containing first and last periods. Ex.: c(2002, 2021).}
}
\value{
A list containing all DCA components
}
\description{
This function runs the entire Data Concentration Analysis to the desired dataset.
A proper explication of the methodology will be soon added do GitHub page.
}
